requirements.txt
absl-py==1.2.0
aiohttp==3.8.1
aiosignal==1.2.0
analytics-python==1.4.0
anyio==3.6.1
asttokens==2.0.8
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
backoff==1.10.0
bcrypt==4.0.0
beautifulsoup4==4.11.1
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==7.1.2
colorama==0.4.5
cryptography==38.0.1
cycler==0.11.0
decorator==5.1.1
executing==1.0.0
fastapi==0.82.0
ffmpy==0.3.0
filelock==3.8.0
flatbuffers==1.12
fonttools==4.37.1
frozenlist==1.3.1
fsspec==2022.8.2
gast==0.4.0
google-auth==2.11.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gradio==3.3
grpcio==1.47.0
h11==0.12.0
h5py==3.7.0
httpcore==0.15.0
httpx==0.23.0
huggingface-hub==0.9.1
idna==3.3
importlib-metadata==4.12.0
ipython==8.4.0
jedi==0.18.1
Jinja2==3.1.2
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
linkify-it-py==1.0.3
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.0
mdurl==0.1.2
monotonic==1.6
multidict==6.0.2
numpy==1.23.2
oauthlib==3.2.0
opt-einsum==3.3.0
orjson==3.8.0
packaging==21.3
pandas==1.4.4
paramiko==2.11.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.30
protobuf==3.19.4
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.15.0
pydantic==1.10.2
pydub==0.25.1
Pygments==2.13.0
PyNaCl==1.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
python-multipart==0.0.5
pytz==2022.2.1
PyYAML==6.0
regex==2022.8.17
requests==2.28.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
sentencepiece==0.1.97
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.5.0
starlette==0.19.1
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
tokenizers==0.12.1
torch==1.12.1
tqdm==4.64.0
traitlets==5.3.0
transformers==4.21.2
typing_extensions==4.3.0
uc-micro-py==1.0.1
urllib3==1.26.12
uvicorn==0.18.3
wcwidth==0.2.5
websockets==10.3
Werkzeug==2.2.2
wikipedia==1.4.0
wrapt==1.14.1
yarl==1.8.1
zipp==3.8.1
